Prune migrated code
In #165715568 [1] we migrated the service broker related code to the
`paas-service-broker-base` repository, because we felt that it made more sense
to be in a place where we could version and maintain it independently.

In this commit, the code which was migrated has been pruned from this
repository, so as not to leave duplicate instances of the code.

[1] https://github.com/alphagov/paas-service-broker-base/pull/1